move requires enroll check to shared mdm folder
