+++ server/service/frontend.go
@@ -1,7 +1,6 @@
 package service
 
 import (
-	"context"
 	"fmt"
 	"html/template"
 	"io"
@@ -11,7 +10,6 @@ import (
 	assetfs "github.com/elazarl/go-bindata-assetfs"
 	shared_mdm "github.com/fleetdm/fleet/v4/pkg/mdm"
 	"github.com/fleetdm/fleet/v4/server/bindata"
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/service/middleware/endpoint_utils"
 	"github.com/go-kit/log"
@@ -123,7 +121,7 @@ func ServeEndUserEnrollOTA(
 			return
 		}
 
-		authRequired, err := RequiresEnrollOTAAuthentication(r.Context(), ds,
+		authRequired, err := shared_mdm.RequiresEnrollOTAAuthentication(r.Context(), ds,
 			enrollSecret, appCfg.MDM.MacOSSetup.EnableEndUserAuthentication)
 		if err != nil {
 			herr(w, "check if authentication is required err: "+err.Error())
@@ -223,36 +221,6 @@ func renderEnrollPage(w io.Writer, appCfg *fleet.AppConfig, urlPrefix, enrollSec
 		return fmt.Errorf("execute react template: %w", err)
 	}
 	return nil
-}
-
-// We take in the AndroidDatastore here, so it can also be called from the android package until https://github.com/fleetdm/fleet/issues/31218 is done
-func RequiresEnrollOTAAuthentication(ctx context.Context, ds fleet.AndroidDatastore, enrollSecret string, noTeamIdPEnabled bool) (bool, error) {
-	secret, err := ds.VerifyEnrollSecret(ctx, enrollSecret)
-	if err != nil && !fleet.IsNotFound(err) {
-		return false, ctxerr.Wrap(ctx, err, "verify enroll secret")
-	}
-
-	if secret == nil {
-		// enroll secret is invalid, check if any team has IdP enabled for setup
-		// experience and if so require authentication before going through (we
-		// enforce the failure due to the enroll secret being invalid only when the
-		// enrollment profile is installed).
-		ids, err := ds.TeamIDsWithSetupExperienceIdPEnabled(ctx)
-		if err != nil {
-			return false, ctxerr.Wrap(ctx, err, "get team IDs with setup experience IdP enabled")
-		}
-		return len(ids) > 0, nil
-	}
-
-	if secret.TeamID == nil { // enroll in "no team"
-		return noTeamIdPEnabled, nil
-	}
-
-	tm, err := ds.Team(ctx, *secret.TeamID)
-	if err != nil {
-		return false, ctxerr.Wrap(ctx, err, "get team for settings")
-	}
-	return tm.Config.MDM.MacOSSetup.EnableEndUserAuthentication, nil
 }
 
 func initiateOTAEnrollSSO(svc fleet.Service, w http.ResponseWriter, r *http.Request, enrollSecret string) error {
